Oceanographic DIC data

CalCOFI collects samples for inorganic carbon and total alkalinity at specific stations throughout the cruise track. Seawater samples are collected from several depths, preserved and brought back to SIO for total inorganic carbon and total alkalinity analysis. These measurements allow the calculation of pH and pCO2. The objectives of these measurements are: first, the long-term characterization of the inorganic carbon system and its response to changing ocean climate, and second, measurements of pH in the coastal zone in order to monitor the anthropogenic impact on coastal ecosystems in the Southern California Bight.

dic

Format

A data frame with 35376 rows and 82 variables

Cst_Cnt

Cast Count - All CalCOFI casts ever conducted, consecutively numbered

Cruise_ID

Cruise identifier Year-Month-Day-C-Ship Code

Cruise.x

Cruise Name Year

Cruz_Sta

Cruise Name and Station Year[Line][Station]

DbSta_ID

Line and Station

Cast_ID

Cast Identifier Century - [YY][MM][ShipCode] - CastType - CastTime-Line

Sta_ID

Line and Station

Quarter

Quarter of the year

Sta_Code

Station Designation (See Station_ID and 0-St_Code for codes)

Distance

Nautical miles from coast intercept, calculated from estimated latitude and longitude

Date

Date (Month Day Year)

Year

Year

Month

Month

Julian_Date

OA Date: days since December 30, 1899

Julian_Day

Julian Day of the year

Time

Time (UTC) CTD reached terminal depth

Lat_Dec

Observed Latitude in decimal degrees

Lat_Deg

Observed Latitude in degrees

Lat_Min

Observed Latitude in minutes

Lat_Hem

Observed Latitude Hemisphere

Lon_Dec

Observed Longitude in decimal degrees

Lon_Deg

Observed Longitude in degrees

Lon_Min

Observed Longitude in minutes

Lon_Hem

Observed Longitude Hemisphere

Rpt_Line

Reported Line Number

St_Line

Nearest Standard Line

Ac_Line

Calculated actual line from observed latitude and longitude

Rpt_Sta

Reported Station Number

St_Station

Nearest Standard Station Number

Ac_Sta

Calculated actual station from observed lat and long

Bottom_D

Bottom depth in meters

Secchi

Secchi disk depth in meters

ForelU

Water color; Only used in CalCOFI dataset from 1988-10 through 1998-04 (Forel-Ule Scale)

Ship_Name

Ship's Name

Ship_Code

Ship's NODC Code

Data_Type

Data Type

Order_Occ

Order station was occupied (within each cruise)

Event_Num

Event number; includes all research opperations within each cruise (e.g. nets, CTD)

Cruz_Leg

Leg of cruise, if multiple legs

Orig_Sta_ID

IEH Reported Station ID

Data_Or

Data Origin, punch card short code used to designate the institution that collected the data (e.g. Scripps)

Cruz_Num

Cruise Designation (Year Month)

IntChl

Integrated Chlorophyll per half light day (milligrams chl per square meter per half light day)

IntC14

Integrated Primary Productivity per half light day (milligrams C per square meter per half light day)

Inc_Str

Incubation Start Time (PST)

Inc_End

Incubation End Time (PST)

PST_LAN

Time of Local Apparent Noon (PST)

Civil_T

Time of Civil Twilight (PST)

TimeZone

Time Zone

Wave_Dir

Wave Direction reported using an abbreviated 360° azimuth circle with 00 representing True North, 18 represents 180°

Wave_Ht

Wave Height in feet

Wave_Prd

Wave Period in seconds

Wind_Dir

Reported using an abbreviated 360 degree azimuth circle with 0 representing True North, 18 representing 180 degrees

Wind_Spd

Wind Speed in knots

Barometer

Millibars to the tenths

Dry_T

Dry Air Temperature from a Sling Psychrometer in degrees Celsius

Wet_T

Wet Air Temperature from a Sling Psychrometer in degrees Celsius

Wea

1 Digit Code from The World Meteorological Organization, Code source WMO 4501

Cloud_Typ

1 Digit Code from The World Meteorological Organization, Code source WMO 0500

Cloud_Amt

1 Digit Code from The World Meteorological Organization, Code source WMO 2700, in oktas

Visibility

1 Digit Code from The World Meteorological Organization, Code source WMO 4300

Cast_Index

CalCOFI Hydrographic Database Cast Table Index; 1 = CalCOFI database's 1st cast record

Btl_Cnt

CalCOFI Hydrographic Database Bottle Table Index; 1 = CalCOFI database's 1st bottle record

Cruise.y

CalCOFI Cruise Designation with century prefix; CalCOFI 0901 = 200901

Depthm

Sample Depth in meters

Depth_ID

Unique database index built from:Century-CruiseShip-Casttype-Julian DOY-UTC Time-LineSta-DepthDup-Record Indicator

DIC1

DIC Analysis - Dissolved Inorganic Carbon Measurement, first sample

DIC2

DIC Analysis - Dissolved Inorganic Carbon Measurement from a duplicate sample drawn from same depth as DIC 1

TA1

DIC Analysis - Total Alkalinity Measurement, first sample; same DIC bottle as DIC 1

TA2

DIC Analysis - Total Alkalinity Measurement from a duplicate sample drawn from same depth as TA 1

pH1

Spectrophotometrically measured pH per kilogram seawater

pH2

Spectrophotometrically measured pH per kilogram seawater from a duplicate sample drawn from same depth as pH 1

Salinity1

DIC Analysis - Salinity value from DIC 1 bottle

Salinity2

DIC Analysis - Salinity value from DIC 2 bottle

Temperature_degC

CTD Temperature from Bottle Sample Depth

Bottle.Salinity

CalCOFI Salinity Measurement from DIC Depth

Bottle_O2_ml_L

CalCOFI Oxygen Measurement in milliliters/liter, from DIC Depth

Bottle_O2_µmol_kg

CalCOFI Oxygen Measurement in umoles/Kg, from DIC Depth

Sigma.theta

Sigma-theta Density calculated from CalCOFI data

DIC.Bottle_ID1

DIC 1 Bottle Index Number

DIC.Bottle_ID2

DIC 2 Bottle Index Number

DIC.Quality.Comment

Quality Control Comments on DIC samples after processing

Source

https://calcofi.org/data/oceanographic-data/dic/

Details

DIC samples for Dave Keeling were also collected on CalCOFI stations 90.70, 90.90, & 90.120 from Sep 1986 (CalCOFI 8609NH) to Nov 2002 (CalCOFI 0211NH). In Aug 2008, DIC sampling restarted on 7 stations, grew to 10 stations in Jan 2011, and currently DIC has expanded substantially. Dissolved inorganic carbon (DIC) and total alkalinity (TA) data have only been integrated into the CalCOFI hydrographic dataset since Aug 2008, then Jan 2009 through presently. Similarly, pH data have only been integrated into the CalCOFI hydrographic dataset since April 2014.

Codes used in data set

Sta_Code (Station Codes)

"ST" - Standard CalCOFI Station

"SCO" - SCCOOS nearshore/20m Station

"NRO" - Not Regularly Occupied Original CalCOFI Station

"OCO" - Occasionally CalCOFI Occupied

"IMX" - IMECOCAL Occupied

"NST" - Non-Standard Station

"MBR" - MBARI Occupied Station

Data_Type (Data Type)

"PR" - Productivity Cast

"HY" - Hydrographic Cast

"10" - Ten-meter Cast

"CT" - Compressed CTD Cast (Low Resolution)

"MX" - Mixed CTD and Bottle Data

_qual, qua, or q (Quality Code; associated with discrete samples; examples: "O_qual", "Chlqua", "PO4q")

Blank - Data OK

"4" - Value zeroed due to value below detection limit

"6" - Data taken from CTD sensor

"8" - Technician thinks value is suspect

"9" - Missing Data

RecInd (Record Indicator)

"3" - Observed Data

"4" - Educated office guess (ghost)

"5" - Data from STD or CTD sensor

"6" - Duplicate Depth

"7" - Interpolated to a standard depth