Plot interactive time series of an oceanographic variable

Plot interactive time series of an oceanographic variable

plot_timeseries(
  d,
  fld_date,
  fld_avg,
  fld_sd,
  title = "Temperature",
  y_label = "Temperature (ºC)",
  color = "red"
)

Arguments

d

data frame with data

fld_date

unquoted name of field containing the date, e.g. year

fld_avg

unquoted name of field containing the average value, e.g. t_deg_c_avg

fld_sd

unquoted name of field containing the standard deviation, e.g. sd

title

quoted title, Default is "Temperature".

y_label

quoted label for y-axis. Default is "Temperature (ºC)".

color

quoted color. Should be one of grDevices::colors(). Default is "red".

Value

interactive plot of dygraphs::dygraph()

Examples

# get variables
(v <- get_variables())
#> Rows: 7 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (5): category, table_field, plot_title, plot_label, plot_color
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 7 × 5
#>   category      table_field                     plot_title plot_label plot_color
#>   <chr>         <chr>                           <chr>      <chr>      <chr>     
#> 1 Oceanographic ctdcast_bottle.t_deg_c          Sea Surfa… Temperatu… red       
#> 2 Oceanographic ctdcast_bottle.salnty           Salinity   Salinity … gray      
#> 3 Oceanographic ctdcast_bottle_dic.bottle_o2_m… Oxygen Co… Oxygen (m… blue      
#> 4 Oceanographic ctdcast_bottle_dic.bottle_o2_m… Oxygen Co… Oxygen (µ… blue      
#> 5 Oceanographic ctdcast_bottle_dic.dic1         Dissolved… DIC (µmol… brown     
#> 6 Oceanographic ctdcast_bottle_dic.dic2         Dissolved… DIC (µmol… brown     
#> 7 Oceanographic ctdcast_bottle.o2sat            Oxygen Sa… Oxygen pe… blue      

# get data for the first variable
(d <- get_timeseries(v$table_field[1]))
#> # A tibble: 71 × 4
#>     year t_deg_c_avg t_deg_c_sd n_obs
#>    <dbl>       <dbl>      <dbl> <dbl>
#>  1  1949        9.12       4.68 28081
#>  2  1950        9.12       4.37 38298
#>  3  1951        9.19       4.54 40459
#>  4  1952       10.4        4.03 31690
#>  5  1953       10.6        3.71 29789
#>  6  1954       10.8        4.33 18874
#>  7  1955       10.1        4.71 23248
#>  8  1956       11.2        4.41 14274
#>  9  1957       13.0        5.04 19072
#> 10  1958       12.1        4.71 24324
#> # … with 61 more rows

# plot time series with the first variable
with(v[1,],
  plot_timeseries(
    # data and columns (from d)
    d, year, t_deg_c_avg, t_deg_c_sd,
    # plot attributes (from v)
    plot_title, plot_label, plot_color))